// src/app/interfaces/productionrecords.ts
export interface ICustomer {
    id: number;
    customerCode: string;
    customerName: string;
    customerShortName: string;
    address1: string;
    address2: string;
    contractPerson: string;
    phoneNo: string;
    faxNo: string;
    emailAddress: string;
    createBy: string;
    createDate: Date;
    updateBy: string;
    updateDate: Date;
    inActivated: boolean;
}
// ...
export interface IProductionOrder {
	productCode: string;
	productName: string;
	orderQty: number;
	fgQty: number;
	deliveryQty: number;
	deliveryMonth: number;
	deliveryYear: number;
	productId: number;
	orderDetail: any[];
}